StockProject_old/backend/requirements.txt
streamlit==1.39.0
requests==2.32.3
PyYAML==6.0.2
pandas==2.2.3
numpy==2.1.1
plotly==5.24.1
matplotlib==3.9.2
python-dateutil==2.9.0
plotly 
pandas
fastapi==0.115.0
uvicorn[standard]==0.30.6
requests==2.32.3
PyYAML==6.0.2
python-multipart==0.0.9
feedparser==6.0.11
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymysql==1.1.0
python-dotenv==1.0.1
pydantic-settings==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0